Breakpad: Require Automake 1.11.1.
Change configure.ac to note that Breakpad requires automake version
1.11.1 or later. This will cause older versions of automake to refuse
to process the Makefile.am file.

Earlier versions of automake generate 'make dist' rules that have a
security flaw; see:

http://web.nvd.nist.gov/view/vuln/detail?vulnId=CVE-2009-4029

However, note that that notice itself has a flaw: the bug is *fixed
in* automake 1.11.1, not present. See:

http://lists.gnu.org/archive/html/automake/2009-12/msg00012.html

(The change to Makefile.in is a consequence of my having neglected to
rebuild Makefile.in after landing r517.)

a=jimblandy, r=mmentovai


git-svn-id: https://google-breakpad.googlecode.com/svn/trunk@521 4c0a9323-5329-0410-9bdc-e9ce6186880e