Fix 'make check' for Android
This patch allows 'make check' to work when performing
an Automake-based build of Breakpad for Android. This
requires to have an Android device connected, and the
'adb' tool in your path.

You can test that with something like:

    configure --host=arm-linux-androideabi
    make check

This is achieved by adding a new small shell script
under android/test-shell.sh, which is invoked by the
Makefile (see TESTS_ENVIRONMENT definition in
Makefile.am).

By default, this runs all unit tests, including those
for the processor and tools (which normally never run
on an Android device).

Note that the test suites fails (e.g. 11 failing tests
for the client library). This will be addressed in later
patches.

+ Modify android/run-checks.sh to run the client library
  test suite on the device by default.

+ Add a new option (--all-tests) to android/run-checks.sh
  which forces it to run the unit test suite for the host
  binaries, and the full suite on the Android device.

+ Update README.ANDROID appropriately.
Review URL: https://breakpad.appspot.com/441002

git-svn-id: https://google-breakpad.googlecode.com/svn/trunk@1023 4c0a9323-5329-0410-9bdc-e9ce6186880e